+++ FirebaseAI/Sources/Extensions/Internal/GeminiLanguageModel+Firebase.swift
@@ -20,9 +20,7 @@
   @available(iOS 27.0, macOS 27.0, watchOS 27.0, visionOS 27.0, *)
   @available(tvOS, unavailable)
   extension GeminiLanguageModel {
-    init(name: String,
-         firebaseAI: FirebaseAI,
-         compatibilityOptions: GeminiLanguageModel.CompatibilityOptions) {
+    init(name: String, firebaseAI: FirebaseAI) {
       let endpointURL = firebaseAI.apiConfig.service.endpoint.rawValue
       guard let urlComponents = URLComponents(string: endpointURL) else {
         preconditionFailure("Invalid Gemini API URL: \(endpointURL)")
@@ -43,8 +41,7 @@
           apiVersion: firebaseAI.apiConfig.version.rawValue
         ),
         headerProvider: firebaseAI.headerProvider,
-        configuration: .ephemeral,
-        compatibilityOptions: compatibilityOptions
+        configuration: .ephemeral
       )
     }
   }