+++ GeminiLanguageModel/Sources/GeminiLanguageModel/GeminiLanguageModel.swift
@@ -53,20 +53,17 @@
     ///   - endpointConfiguration: The network endpoint configuration.
     ///   - headerProvider: An optional async provider for dynamic headers (such as auth tokens).
     ///   - configuration: The `URLSessionConfiguration` to use. Defaults to `.ephemeral`.
-    ///   - compatibilityOptions: Options for configuring compatibility with the Gemini API.
     package init(
       modelResource: ModelResource,
       endpointConfiguration: EndpointConfiguration,
       headerProvider: (@Sendable () async throws -> [String: String])? = nil,
-      configuration: URLSessionConfiguration = .ephemeral,
-      compatibilityOptions: CompatibilityOptions
+      configuration: URLSessionConfiguration = .ephemeral
     ) {
       executorConfiguration = Executor.Configuration(
         modelResource: modelResource,
         endpointConfiguration: endpointConfiguration,
         headerProvider: headerProvider.map { HeaderProvider($0) },
-        sessionConfiguration: configuration,
-        compatibilityOptions: compatibilityOptions
+        sessionConfiguration: configuration
       )
     }
   }